refactor(j-s): move ModifyDatesModal to seprate component
